PASSWORD GENERATOR
A PASSWORD GENERATOR is a useful tool that generate strong and random password for users. This project aims to create a password generator application using python , allowing users to specify the length and complexity of the password .

USER INPUT : Prompt the user to specify the desire length of the password,

GENERATE PASSWORD : Use a combination of random characters to generate a password  of the specified length.

DISPLAY THE PASSWORD : Print the generated password on the screen.